+++ resources/core/cljs/home.cljs
@@ -5,7 +5,8 @@
 {{=<< >>=}}
 (ns <<namespace>>.client.home
   (:require [re-frame.core :as rf]<<#hydrogen-session?>>
-            [<<namespace>>.client.session :as session]<</hydrogen-session?>>
+            [<<namespace>>.client.session :as session]
+            [<<namespace>>.client.user :as user]<</hydrogen-session?>>
             [<<namespace>>.client.view :as view]))
 
 (rf/reg-event-fx
@@ -14,7 +15,16 @@
    {:dispatch [::view/set-active-view :home]
     :redirect "/#/home"}))<<#hydrogen-session?>>
 
-(defn logout []
+(defn- user-details []
+  (let [user-data (rf/subscribe [::user/user-data])]
+    (fn []
+      (when @user-data
+        [:div {:style {:text-align :center}}
+         [:p (str "Hello again " (:first-name @user-data) " " (:last-name @user-data) "!")]
+         [:img {:src (or (:avatar @user-data) "images/user.svg")
+                :style {:width "100px" :height "100px" :border-radius "50%"}}]]))))
+
+(defn- logout []
   [:div.logout
    {:on-click #(do (rf/dispatch [::session/user-logout])
                    (view/redirect! "/#/landing"))}
@@ -24,4 +34,5 @@
   [:div {:id "home"}
    [:img {:src "images/hydrogen-logo-white.svg" :alt "Hydrogen logo"}]
    [:h1 "Welcome to Hydrogen!"]<<#hydrogen-session?>>
+   [user-details]
    [logout]<</hydrogen-session?>>])