+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Changed
+- Stopped using cookies for tracking keycloak process status.
 
 ## [0.3.2] - 2019-11-03
 ### Changed