+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+
+## [0.3.3] - 2019-12-05
 ### Added
 - [Keycloak] - User data is fetched on successful login and displayed in home ns.
 - Tooltip, generic popup and loading popup best practices.
@@ -94,8 +96,9 @@ were moved to an index of suggested next steps in the relevant `.md` file.
 Sometimes it was `re-frame` and sometimes it was `rf`.
 This version unifies it to `rf`.
  
-[UNRELEASED]:  https://github.com/magnetcoop/hydrogen.duct-template/compare/v0.3.2...HEAD
-[0.3.1]: https://github.com/magnetcoop/hydrogen.duct-template/compare/v0.3.1...v0.3.2
+[UNRELEASED]:  https://github.com/magnetcoop/hydrogen.duct-template/compare/v0.3.3...HEAD
+[0.3.3]: https://github.com/magnetcoop/hydrogen.duct-template/compare/v0.3.2...v0.3.3
+[0.3.2]: https://github.com/magnetcoop/hydrogen.duct-template/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/magnetcoop/hydrogen.duct-template/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/magnetcoop/hydrogen.duct-template/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/magnetcoop/hydrogen.duct-template/compare/v0.1.7...v0.2.0